# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Videogames API",
        "description": "API REST para gestión de videojuegos y plataformas con autenticación OAuth 2.0",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local de Desarrollo"
        }
    ],
    "paths": {
...
        "/api/v1/plataformas": {
            "get": {
                "tags": [
                    "Plataformas"
                ],
                "summary": "Listar todas las plataformas",
                "description": "Obtiene la lista completa de plataformas con el conteo de videojuegos",
                "operationId": "2cfbc3841b923db4746a20e5de6d0d35",
                "responses": {
                    "200": {
                        "description": "Lista de plataformas obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "PlayStation 5"
                                                    },
                                                    "fabricante": {
                                                        "type": "string",
                                                        "example": "Sony"
                                                    },
                                                    "videojuegos_count": {
                                                        "type": "integer",
                                                        "example": 5
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Plataformas"
                ],
                "summary": "Crear plataforma (Solo Admin)",
                "description": "Crea una nueva plataforma. Requiere permisos de administrador",
                "operationId": "4dae9d177dd781d734f220ca983ea04f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Steam Deck"
                                    },
                                    "fabricante": {
                                        "type": "string",
                                        "example": "Valve"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plataforma creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plataforma creada exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 7
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Steam Deck"
                                                },
                                                "fabricante": {
                                                    "type": "string",
                                                    "example": "Valve"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No tienes permisos (solo admin)"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/plataformas/{id}": {
            "get": {
                "tags": [
                    "Plataformas"
                ],
                "summary": "Ver plataforma por ID",
                "description": "Obtiene una plataforma específica con sus videojuegos",
                "operationId": "8695b3b71ea5682dc5cf9f6613990929",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la plataforma",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plataforma encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "PlayStation 5"
                                                },
                                                "fabricante": {
                                                    "type": "string",
                                                    "example": "Sony"
                                                },
                                                "videojuegos": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "titulo": {
                                                                "type": "string"
                                                            },
                                                            "anio_lanzamiento": {
                                                                "type": "string",
                                                                "format": "date"
                                                            },
                                                            "genero": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plataforma no encontrada"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Plataformas"
                ],
                "summary": "Actualizar plataforma (Solo Admin)",
                "description": "Actualiza una plataforma existente. Requiere permisos de administrador",
                "operationId": "4273b760b15e96618bc243e756520c93",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la plataforma",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Steam Deck OLED"
                                    },
                                    "fabricante": {
                                        "type": "string",
                                        "example": "Valve Corporation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plataforma actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plataforma actualizada exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "nombre": {
                                                    "type": "string"
                                                },
                                                "fabricante": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No tienes permisos"
                    },
                    "404": {
                        "description": "Plataforma no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Plataformas"
                ],
                "summary": "Eliminar plataforma (Solo Admin)",
                "description": "Elimina una plataforma. Requiere permisos de administrador",
                "operationId": "b597a978a56a7bef233e7cebb079ab67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la plataforma",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plataforma eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plataforma eliminada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No tienes permisos"
                    },
                    "404": {
                        "description": "Plataforma no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/plataformas/mas-popular": {
            "get": {
                "tags": [
                    "Plataformas"
                ],
                "summary": "Obtener plataforma más popular",
                "description": "Obtiene la plataforma con mayor cantidad de videojuegos",
                "operationId": "b50c2cb10a4e2ac741281ee521a66c0d",
                "responses": {
                    "200": {
                        "description": "Plataforma más popular obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "PC"
                                                },
                                                "fabricante": {
                                                    "type": "string",
                                                    "example": "Varios"
                                                },
                                                "videojuegos_count": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "videojuegos": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "titulo": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "La plataforma más popular es PC con 10 videojuegos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay plataformas disponibles"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa tu token Bearer obtenido del endpoint de login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}